US138560 - refactored multi-select
